requirements.txt
django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
pyjwt==2.9.0 